Clean up color system
